Move settings.h to setup
