+++ src/eip4844/api.c
@@ -20,8 +20,8 @@
 #include "common/fr.h"
 #include "common/lincomb.h"
 #include "common/ret.h"
-#include "common/settings.h"
 #include "common/utils.h"
+#include "setup/settings.h"
 
 #include <assert.h> /* For assert */
 #include <stdlib.h> /* For NULL */
